add LoginCookieProvider api
